Pull functions from vacancy.js route file into vacancy module
