+++ lib/modules/vacancy/index.js
@@ -8,6 +8,11 @@ const fetchVacancyList = async function(filters) {
     return await response.json();
 };
 
+const fetchVacancyDetails = async function(id) {
+    let response = await fetch(`${process.env.API_URL}:${process.env.API_PORT}/vacancy/${id}`);
+    return await response.json();
+};
+
 const formatResultData = (results = []) =>
     results.map(result => {
         const salary = formatSalaryOutput(result.salaryMin, result.salaryMax);
@@ -34,9 +39,18 @@ const formatSalaryOutput = (min, max) => {
 
 const formatSalaryNumber = num => `£${num.toLocaleString()}`;
 
+const generateReturnURL = queryStr => {
+    if(!queryStr) return '/';
+
+    const qs = queryStr.replace(/return_url=/, '');
+    return `/results?${qs}`;
+};
+
 module.exports = {
     fetchVacancyList,
+    fetchVacancyDetails,
     formatResultData,
     formatSalaryOutput,
-    formatSalaryNumber
+    formatSalaryNumber,
+    generateReturnURL
 };
